+++ aws-java-sdk-redshift/src/main/java/com/amazonaws/services/redshift/model/transform/DescribeResizeResultStaxUnmarshaller.java
@@ -129,6 +129,11 @@ public class DescribeResizeResultStaxUnmarshaller implements Unmarshaller<Descri
                     describeResizeResult.setMessage(StringStaxUnmarshaller.getInstance().unmarshall(context));
                     continue;
                 }
+
+                if (context.testExpression("TargetEncryptionType", targetDepth)) {
+                    describeResizeResult.setTargetEncryptionType(StringStaxUnmarshaller.getInstance().unmarshall(context));
+                    continue;
+                }
             } else if (xmlEvent.isEndElement()) {
                 if (context.getCurrentDepth() < originalDepth) {
                     return describeResizeResult;