+++ aws-java-sdk-redshift/src/main/java/com/amazonaws/services/redshift/model/transform/PendingModifiedValuesStaxUnmarshaller.java
@@ -92,6 +92,11 @@ public class PendingModifiedValuesStaxUnmarshaller implements Unmarshaller<Pendi
                     pendingModifiedValues.setMaintenanceTrackName(StringStaxUnmarshaller.getInstance().unmarshall(context));
                     continue;
                 }
+
+                if (context.testExpression("EncryptionType", targetDepth)) {
+                    pendingModifiedValues.setEncryptionType(StringStaxUnmarshaller.getInstance().unmarshall(context));
+                    continue;
+                }
             } else if (xmlEvent.isEndElement()) {
                 if (context.getCurrentDepth() < originalDepth) {
                     return pendingModifiedValues;