+++ aws-java-sdk-route53/src/main/java/com/amazonaws/services/route53domains/model/transform/UpdateDomainContactRequestProtocolMarshaller.java
@@ -22,6 +22,7 @@ import com.amazonaws.services.route53domains.model.*;
 import com.amazonaws.transform.Marshaller;
 
 import com.amazonaws.protocol.*;
+import com.amazonaws.protocol.Protocol;
 import com.amazonaws.annotation.SdkInternalApi;
 
 /**