+++ aws-java-sdk-xray/src/main/java/com/amazonaws/services/xray/model/transform/UpdateSamplingRuleRequestProtocolMarshaller.java
@@ -22,6 +22,7 @@ import com.amazonaws.services.xray.model.*;
 import com.amazonaws.transform.Marshaller;
 
 import com.amazonaws.protocol.*;
+import com.amazonaws.protocol.Protocol;
 import com.amazonaws.annotation.SdkInternalApi;
 
 /**